# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Shop Api",
        "version": "1.0.0"
    },
    "host": "shop.loc",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products with images and pagination",
                "parameters": [
                    {
                        "name": "pagination",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "minimum": 1
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "minimum": 1
                    },
                    {
                        "name": "currency_type",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "usd",
                            "eur",
                            "uah"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "current_page": {
                                    "type": "integer"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "allOf": [
                                            {
                                                "$ref": "#/definitions/Product"
                                            },
                                            {
                                                "properties": {
                                                    "product_images": {
                                                        "type": "array",
                                                        "items": {
                                                            "$ref": "#/definitions/ProductImage"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        ]
                                    }
                                },
                                "first_page_url": {
                                    "type": "string"
                                },
                                "from": {
                                    "type": "integer"
                                },
                                "last_page": {
                                    "type": "integer"
                                },
                                "last_page_url": {
                                    "type": "string"
                                },
                                "next_page_url": {
                                    "type": "string"
                                },
                                "path": {
                                    "type": "string"
                                },
                                "per_page": {
                                    "type": "integer"
                                },
                                "prev_page_url": {
                                    "type": "string"
                                },
                                "to": {
                                    "type": "integer"
                                },
                                "total": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "operationId": "createProduct",
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "currency",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "usd",
                            "eur",
                            "uah"
                        ]
                    },
                    {
                        "name": "images[0]",
                        "in": "formData",
                        "description": "request can attempt array images[i]:{file} /mimes:jpeg,bmp,png/max size:2Mb",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "images[1]",
                        "in": "formData",
                        "description": "this example additing more images",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "categories[0]",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products with images and pagination",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "integer",
                        "minimum": 1
                    },
                    {
                        "name": "currency_type",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "usd",
                            "eur",
                            "uah"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Product"
                                },
                                {
                                    "properties": {
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/definitions/Category"
                                                    },
                                                    {
                                                        "properties": {
                                                            "pivot": {
                                                                "properties": {
                                                                    "product_id": {
                                                                        "type": "integer"
                                                                    },
                                                                    "category_id": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                ]
                                            }
                                        },
                                        "product_images": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ProductImage"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of product",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id",
                        "in": "formData",
                        "description": "Need to validate for update",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "currency",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "usd",
                            "eur",
                            "uah"
                        ]
                    },
                    {
                        "name": "categories[0]",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Product is not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Remove product",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "id number of product",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Product is not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Category": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "create_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            },
            "type": "object"
        },
        "Product": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            },
            "type": "object"
        },
        "ProductImage": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "link_to_file": {
                    "type": "string"
                },
                "link_to_thumb": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            },
            "type": "object"
        }
    }
}